LayoutTests/platform/ios/fast/events/ios/key-command-select-all-prevent-default-expected.txt

Tests that preventDefault() when Command + A is pressed prevents selection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getSelection().type is "Caret"
PASS successfullyParsed is true

TEST COMPLETE
